+++ jwst/assign_wcs/tests/test_wcs.py
@@ -248,7 +248,9 @@ def test_sip_approx(tmp_path):
     im = ImageModel(hdu1)
 
     pipe = AssignWcsStep()
-    result = pipe.call(im)
+    result = pipe.call(im, sip_max_pix_error=0.1, sip_degree=3,
+                       sip_max_inv_pix_error=0.1, sip_inv_degree=3,
+                       sip_npoints=12)
 
     # check that result.meta.wcsinfo has correct
     # values after SIP approx.
@@ -278,4 +280,4 @@ def test_sip_approx(tmp_path):
     result.write(path)
 
     with open(path) as result_read:
-        result.meta.wcsinfo == result_read.meta.wcsinfo
+        assert result.meta.wcsinfo == result_read.meta.wcsinfo